+++ java/hibernate/dialect/src/test/java/software/amazon/dsql/hibernate/dialect/AuroraDSQLDialectTest.java
@@ -55,8 +55,8 @@ public class AuroraDSQLDialectTest {
   public void testForeignKeySupport() {
     Exporter<ForeignKey> exporter = dialect.getForeignKeyExporter();
     assertNotNull(exporter);
-    assertFalse(dialect.supportsCascadeDelete());
-    assertFalse(dialect.dropConstraints());
+    assertTrue(dialect.supportsCascadeDelete());
+    assertTrue(dialect.dropConstraints());
   }
 
   @Test